Zoom and moving the camera now works with SDL_FINGER* events
--HG--
branch : hedgeroid

+++ project_files/Android-build/SDL-android-project/src/org/hedgewars/mobile/TouchInterface/TouchInterface.java
@@ -40,7 +40,8 @@ class TouchInterfaceMT implements OnTouchListener {
 		int actionCode = action & MotionEvent.ACTION_MASK;
 		
 		for (int i = 0; i < event.getPointerCount(); i++) {
-			SDLActivity.onNativeTouch(actionCode, event.getPointerId(i), event.getX(i), event.getY(i), event.getPressure(i));
+			SDLActivity.onNativeTouch(actionCode, event.getPointerId(i), (int)event.getX(i), (int)event.getY(i), event.getPressure(i));
+//			Log.d("Android", String.format("x=%f, y=%f, pntr=%d", event.getX(i), event.getY(i), event.getPointerId(i)));
 		   }
 		return true;
 	}
